feat: Phase 2 connect-demo controls, PRNDL selector, torque telemetry, twin feedback loop
Add PRNDL gear selector (P/R/N/D) with keyboard cycling via [ and ],
clamped at boundaries. Engine idles in PARK like NEUTRAL (no spurious
cranking). Throttle holds for 9 frames (~100ms) to bridge OS key repeat
gaps, then snaps to zero on release. Ignition toggle via 'i' key.

ZF 8HP45 gearbox now uses actual engine-sim vehicle speed feedback
instead of DemoVehiclePhysics model, fixing erratic shift decisions.
Torque display reads clutch constraint F_t values directly, showing
engine-side and gear-multiplied drivetrain-side torque separately.

Also: out-of-source build guard, -Werror on bridge targets only,
ManualTwin pure passthrough, component architecture diagram (mermaid),
Phase 2 acceptance criteria (86 ACs), ZF torque-aware shifting research.

Update engine-sim submodule pointer for transmission torque accessors.

+++ src/common/CircularBuffer.cpp
@@ -66,7 +66,7 @@ size_t CircularBuffer::write(const float* samples, size_t frameCount) {
     int writePtr = writePointer_.load();
 
     // Handle wrap-around
-    if (writePtr + toWrite <= static_cast<int>(bufferCapacity_)) {
+    if (writePtr + static_cast<int>(toWrite) <= static_cast<int>(bufferCapacity_)) {
         // Simple case: no wrap-around
         for (size_t i = 0; i < toWrite; i++) {
             buffer_[(writePtr + i) * 2] = samples[i * 2];
@@ -112,7 +112,7 @@ size_t CircularBuffer::read(float* output, size_t frameCount) {
     int readPtr = readPointer_.load();
 
     // Handle wrap-around
-    if (readPtr + toRead <= static_cast<int>(bufferCapacity_)) {
+    if (readPtr + static_cast<int>(toRead) <= static_cast<int>(bufferCapacity_)) {
         // Simple case: no wrap-around
         for (size_t i = 0; i < toRead; i++) {
             output[i * 2] = buffer_[(readPtr + i) * 2];
